requirements.txt
opencv-python>=4.8
flask>=3.0
flask-cors>=4.0
python-dotenv>=1.0
prometheus-client>=0.19
werkzeug>=3.0
scipy>=1.11
mediapipe>=0.10
